capdl: Consolidate initializer into one package
