--- crates/sel4-capdl-initializer/core/Cargo.toml
@@ -1,23 +1,0 @@
-#
-# Copyright 2023, Colias Group, LLC
-#
-# SPDX-License-Identifier: BSD-2-Clause
-#
-#
-# This file is generated from './Cargo.nix'. You can edit this file directly
-# if you are not using this project's Cargo manifest management tools.
-# See 'hacking/cargo-manifest-management/README.md' for more information.
-#
-
-[package]
-name = "sel4-capdl-initializer-core"
-version = "0.1.0"
-authors = ["Nick Spinale <nick.spinale@coliasgroup.com>"]
-edition = "2024"
-license = "BSD-2-Clause"
-
-[dependencies]
-log = "0.4.28"
-rkyv = { version = "0.8.12", default-features = false }
-sel4 = { path = "../../sel4" }
-sel4-capdl-initializer-types = { path = "../types", features = ["sel4"] }
