+++ crates/sel4-capdl-initializer/src/lib.rs
@@ -6,16 +6,19 @@
 
 #![no_std]
 
+#[cfg(feature = "alloc")]
+extern crate alloc;
+
 mod buffers;
 mod cslot_allocator;
 mod error;
 mod hold_slots;
 mod initialize;
+mod lib_main;
 mod memory;
 
-pub use buffers::{InitializerBuffers, PerObjectBuffer};
-pub use error::CapDLInitializerError;
-pub use initialize::Initializer;
+#[cfg(not(feature = "alloc"))]
+mod no_allocator;
 
 #[sel4::sel4_cfg(all(ARCH_RISCV64, not(PT_LEVELS = "3")))]
 compile_error!("unsupported configuration");
