+++ crates/sel4-capdl-initializer/src/memory.rs
